Added tokens + user auth in backend and frontend  - SignUp
Now we are storing hashed passwords. The sign up process linked to the dashboard is now secure. User Authentication has been done on this route for now. Has been taken care of in both students and managers.

Used bcryptjs, jsonwebtoken, and dotenv.

A Middleware file was also added for the Auth purpose + an .env file was also added for security.