Included User Auth + Token in Student Login
Just the same process, but repeated for Login. Now the sign up firstly leads to the login page and then towards the dashboard, where you can log out as well.

The StudentLogin Procedure is not being used now. Instead we get the Password using the email and then compare it with the input one using bcrypt (hashing and salts).