Included User Auth + Token in Manager Login
Same stuff as Student Login. Have only done so for PManager and RManager. GManager would be handled after the Grounds update. Login and Sign up are seamless now. Hashed passwords + already existing ones such as photocopier and ground are handled with generated hashed passwords.

authenticateManager was added to the middleware file. This will be used for authenticating if they have the role required to do the work.