Added Authorization + Further Authentication
In addition to authentication, authorization added at backend in middleware. Checks for role of user. Updated entire frontend (all pages except grounds) for authentication token when sending requests.

General Function by the name of authorizedFetch() included for reusability. It can be further generalized as an imported function which would work in all files according to the roles.