Added Student + Manager Views
Controls the Active and Inactive State of the Students. Can delete the inactive students. The New Students which sign up are inactive at first and need to be activated by the admin. Active state checked at login.

Managers can be deleted except for the ID: 1 and 2. Requires admin's password for confirmation.

SQL File updated for deletion of managers. Restaurants and Operational Timings deleted on cascade through manager deletion.